Add multiple OSRM server options and custom URL functionality
The commit introduces selectbox functionality for choosing the OSRM server to use for routing calculations. As part of this, a text input allows for a custom server URL when 'Custom' is selected from the options. These enhancements will give the user a wider choice of routing servers, including setting up their own routing server if desired. A mapping between transport modes and OSRM profiles has also been implemented when using a custom server.