Add method to get all OSRM profiles and optimize trip calculation
A new method, `get_all_osrm_profiles`, has been added for accessing all OSRM profiles. In addition, the `calculate_trip` function has been refactored to incorporate an optimization check. If the `optimize_points` flag is set, the function performs the join with the street network, otherwise it is not necessary.